# perf: improve SaveMetaInfoToMap performance