## Benchmark
**Before**:
```
BenchmarkAllParallel/SaveMetaInfoToMap_10
BenchmarkAllParallel/SaveMetaInfoToMap_10-12         	  912304	      1723 ns/op	    3720 B/op	      28 allocs/op
BenchmarkAllParallel/SaveMetaInfoToMap_20
BenchmarkAllParallel/SaveMetaInfoToMap_20-12         	  530864	      2380 ns/op	    8026 B/op	      50 allocs/op
BenchmarkAllParallel/SaveMetaInfoToMap_50
BenchmarkAllParallel/SaveMetaInfoToMap_50-12         	  205908	      5524 ns/op	   18170 B/op	     118 allocs/op
BenchmarkAllParallel/SaveMetaInfoToMap_100
BenchmarkAllParallel/SaveMetaInfoToMap_100-12        	   87831	     14197 ns/op	   36655 B/op	     226 allocs/op
```

**After**:
```
BenchmarkAllParallel/SaveMetaInfoToMap_10
BenchmarkAllParallel/SaveMetaInfoToMap_10-12         	 1258407	       897.6 ns/op	    2331 B/op	      22 allocs/op
BenchmarkAllParallel/SaveMetaInfoToMap_20
BenchmarkAllParallel/SaveMetaInfoToMap_20-12         	  639684	      1971 ns/op	    5404 B/op	      44 allocs/op
BenchmarkAllParallel/SaveMetaInfoToMap_50
BenchmarkAllParallel/SaveMetaInfoToMap_50-12         	  273380	      4734 ns/op	   12570 B/op	     109 allocs/op
BenchmarkAllParallel/SaveMetaInfoToMap_100
BenchmarkAllParallel/SaveMetaInfoToMap_100-12        	   98899	     11902 ns/op	   25629 B/op	     214 allocs/op
```